+++ backend/logic/homebrewery.ts
@@ -1,11 +1,11 @@
+import { checkBestiaryPermission } from "./bestiaries";
+import { checkCreaturePermission } from "./creatures";
+import { possibleUser } from "./login";
 import type { FeatureEntity, SaveEntity, Stat, Statblock } from "~/shared";
 import { app } from "@/utilities/constants";
 import { getBestiary, getCreature, getCreaturesByBestiary, incrementBestiaryViewCount } from "@/utilities/database";
 import { log } from "@/utilities/logger";
-import { capitalizeFirstLetter, displayCasterCasting, displayInnateCasting, displaySpeedOrSenses, hpCalc, ppCalc, SKILLS_BY_STAT, statCalc } from "~/shared";
-import { checkBestiaryPermission } from "./bestiaries";
-import { checkCreaturePermission } from "./creatures";
-import { possibleUser } from "./login";
+import { SKILLS_BY_STAT, capitalizeFirstLetter, displayCasterCasting, displayInnateCasting, displaySpeedOrSenses, hpCalc, ppCalc, statCalc } from "~/shared";
 
 app.get("/api/homebrewery/export/bestiary/:id", possibleUser, async (req, res) => {
 	try {