+++ frontend/src/components/VisualEditor/Nodes/NodeText.vue
@@ -29,7 +29,7 @@ watch(() => descIsText.value, () => {
 
 useDataCleanup(currentEffect, ["title"]);
 
-const setDesc = inject<any>("setActionDescription");
+const setDesc = inject<false | Function>("setActionDescription");
 </script>
 
 <template>
@@ -44,7 +44,7 @@ const setDesc = inject<any>("setActionDescription");
 		<LabelledComponent title="Description" for="text" style="margin-top: 1rem">
 			<div v-if="descIsText" class="input-wrapper">
 				<textarea id="text" v-model="(currentEffect.text as string)" rows="5" placeholder="Description" /><AnnotatedString />
-				<small style="font-size: x-small; cursor: pointer" role="button" @click="setDesc(currentEffect.text)"> <i>Set the description of the statblock trait to this text.</i> </small>
+				<small v-if="setDesc" style="font-size: x-small; cursor: pointer" role="button" @click="setDesc(currentEffect.text)"> <i>Set the description of the statblock trait to this text.</i> </small>
 			</div>
 			<v-select v-else v-model="currentEffect.text" :options="abilities" label="name" input-id="text" :reduce="(x : any) => ({ id: x.id, typeId: x.typeId })" />
 		</labelledcomponent>