+++ frontend/src/views/BestiaryEdit.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import type { Bestiary, BestiaryExtended, CreatureWithStats, Statblock, User } from "~/shared";
 import { createPopper } from "@popperjs/core";
 import { Shimmer } from "@shimmer-from-structure/vue";
 import { refDebounced, useLocalStorage } from "@vueuse/core";
 import { onMounted, ref, watch } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import Draggable from "vuedraggable";
+import type { Bestiary, BestiaryExtended, CreatureWithStats, Statblock, User } from "~/shared";
 import CopyManager from "@/components/Bestiary/CopyManager.vue";
 import CreatureListItem from "@/components/Bestiary/CreatureListItem.vue";
 import StatusIcon from "@/components/Bestiary/StatusIcon.vue";
@@ -646,7 +646,7 @@ const getDraggableKey = (item: any) => {
 			v-if="bestiary"
 			:routes="[
 				{
-					path: isOwner || isEditor ? '../my-bestiaries/' : '../bestiaries',
+					path: isOwner || isEditor ? '/bestiaries/personal' : '/bestiaries/public',
 					text: isOwner || isEditor ? 'My Bestiaries' : 'Bestiaries',
 					isCurrent: false
 				},